Add server wrapper for waiting room page
